Split between internal and external views
